Implement cleanup node updates
