+++ ckworker/ckworker/ckcore.py
@@ -98,24 +98,6 @@ def send_graph(
 
 def add_args(arg_parser: ArgumentParser) -> None:
     arg_parser.add_argument(
-        "--ckcore-uri",
-        help="ckcore URI (default: http://localhost:8900)",
-        default="http://localhost:8900",
-        dest="ckcore_uri",
-    )
-    arg_parser.add_argument(
-        "--ckcore-ws-uri",
-        help="ckcore Websocket URI (default: ws://localhost:8900)",
-        default="ws://localhost:8900",
-        dest="ckcore_ws_uri",
-    )
-    arg_parser.add_argument(
-        "--ckcore-graph",
-        help="ckcore graph name (default: ck)",
-        default="ck",
-        dest="ckcore_graph",
-    )
-    arg_parser.add_argument(
         "--debug-dump-json",
         help="Dump the generated json data (default: False)",
         dest="debug_dump_json",
